UI Clean Up: re-factor Grading, Grading+, and Gradebook pages into panels; Streamline InstructorDashboard page further
