+++ src/renderer/src/assets/main.css
@@ -70,7 +70,7 @@ footer {
 }
 
 .versions {
-  position: absolute;
+  position: fixed;
   bottom: 30px;
   margin: 0 auto;
   padding: 15px 0;
@@ -954,7 +954,36 @@ footer {
 .dashboard-toolbar {
   display: flex;
   gap: 1rem;
+  flex-wrap: wrap;
   margin-top: 1.5rem;
+}
+
+.dashboard-tab-button.active {
+  background-color: var(--ev-c-blue-1);
+  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.18);
+  transform: translateY(-1px);
+}
+
+.dashboard-tab-button.active:hover {
+  background-color: var(--ev-c-blue-1);
+  transform: translateY(-1px);
+}
+
+.dashboard-panel-transition {
+  animation: dashboard-panel-reveal 220ms ease-out;
+  transform-origin: top center;
+}
+
+@keyframes dashboard-panel-reveal {
+  from {
+    opacity: 0;
+    transform: translateY(8px) scale(0.995);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
 }
 
 .dashboard-empty-state {
